split MailSender, etc. into separate implementations for MailKit and System.Net.Mail
