+++ Directory.Build.props
@@ -1,0 +1,17 @@
+<Project>
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+
+        <Nullable>enable</Nullable>
+
+        <Version>1.2.0</Version>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <Authors>TRENZ GmbH</Authors>
+        <PackageProjectUrl>https://github.com/trenz-gmbh/TRENZ.Lib.RazorMail</PackageProjectUrl>
+        <License>MIT</License>
+        <RepositoryUrl>git@github.com:trenz-gmbh/TRENZ.Lib.RazorMail.git</RepositoryUrl>
+        <RepositoryType>git</RepositoryType>
+    </PropertyGroup>
+</Project>
