+++ TRENZ.Lib.RazorMail.SampleWebApi/TRENZ.Lib.RazorMail.SampleWebApi.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\TRENZ.Lib.RazorMail.MailKit\TRENZ.Lib.RazorMail.MailKit.csproj" />
       <ProjectReference Include="..\TRENZ.Lib.RazorMail\TRENZ.Lib.RazorMail.csproj" />
 <!--        <PackageReference Include="TRENZ.Lib.RazorMail" Version="1.0.0" />-->
     </ItemGroup>
