+++ TRENZ.Lib.RazorMail/Models/MailAttachment.cs
@@ -1,7 +1,5 @@
 ﻿using System.IO;
 
-using MimeKit;
-
 namespace TRENZ.Lib.RazorMail.Models;
 
 public class MailAttachment
@@ -19,33 +17,4 @@ public class MailAttachment
 
     public string? ContentId { get; set; }
     public bool Inline { get; set; }
-
-    public static implicit operator System.Net.Mail.Attachment(MailAttachment attachment)
-    {
-        var result = new System.Net.Mail.Attachment(new MemoryStream(attachment.FileData), attachment.Filename,
-            attachment.ContentType);
-
-        result.ContentId = attachment.ContentId;
-        result.ContentDisposition.Inline = attachment.Inline;
-
-        return result;
-    }
-
-    public static implicit operator MimePart(MailAttachment attachment)
-    {
-        var mimeContent = new MimeContent(new MemoryStream(attachment.FileData));
-        var contentDisposition =
-            new ContentDisposition(attachment.Inline ? ContentDisposition.Inline : ContentDisposition.Attachment);
-
-        var result = new MimePart(MimeKit.ContentType.Parse(attachment.ContentType))
-        {
-            ContentId = attachment.ContentId,
-            Content = mimeContent,
-            ContentDisposition = contentDisposition,
-            ContentTransferEncoding = ContentEncoding.Base64,
-            FileName = attachment.Filename
-        };
-
-        return result;
-    }
 }
