+++ TRENZ.Lib.RazorMail/TRENZ.Lib.RazorMail.csproj
@@ -1,21 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
     <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
 
-    <Nullable>enable</Nullable>
     <RootNamespace>TRENZ.Lib.RazorMail</RootNamespace>
-
-    <Version>1.1.0</Version>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Authors>TRENZ GmbH</Authors>
-    <PackageProjectUrl>https://github.com/trenz-gmbh/TRENZ.Lib.RazorMail</PackageProjectUrl>
-    <License>MIT</License>
-    <RepositoryUrl>git@github.com:trenz-gmbh/TRENZ.Lib.RazorMail.git</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
   </PropertyGroup>
 
   <ItemGroup>
@@ -23,7 +11,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="MailKit" Version="2.2.0" />
     <PackageReference Include="NUnit" Version="3.13.2" />
     <PackageReference Include="NLog" Version="4.7.11" />
   </ItemGroup>
